refactor(movie): 어댑터명 변경
